Code_RL/requirements.txt
# Core RL and ML libraries
gymnasium>=0.29.0
stable-baselines3>=2.0.0
torch>=2.0.0
numpy>=1.21.0
matplotlib>=3.5.0
pandas>=1.3.0

# Configuration and utilities
pyyaml>=6.0
pydantic>=2.0.0
dataclasses-json>=0.5.0
click>=8.0.0

# Testing and quality
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0

# Networking (if using HTTP endpoint)
requests>=2.28.0
httpx>=0.24.0
aiohttp>=3.8.0
urllib3>=1.26.0
websockets>=11.0.0

# Performance optimization and monitoring
tenacity>=8.0.0
prometheus_client>=0.16.0
asyncio-throttle>=1.0.2

# Compression support
brotli>=1.0.0
lz4>=4.0.0

# Additional utilities for fault tolerance
python-dateutil>=2.8.0
psutil>=5.9.0

# Data analysis and visualization
seaborn>=0.11.0
plotly>=5.0.0
jupyter>=1.0.0

# Optional: ARZ simulator dependencies
scipy>=1.9.0
numba>=0.56.0
